refactor(reservation): 컨트롤러를 reservation_time 구조에 맞게 재작성
